Expand map section ID type to u16

+++ src/region_map.c
@@ -2100,7 +2100,7 @@ bool8 IsRegionMapZoomed(void)
     return sRegionMap->zoomed;
 }
 
-bool32 IsEventIslandMapSecId(u8 mapSecId)
+bool32 IsEventIslandMapSecId(u16 mapSecId)
 {
     u32 i;
